requirements.txt
streamlit>=1.32.0
pandas>=2.1.0
groq>=0.4.0
python-dotenv>=1.0.0
